requirements.txt
# ─────────────────────────────────────────────────────────────
# Semantic Halting Problem (SHP) — Python Dependencies
#
# Install:  pip install -r requirements.txt
# Tested with Python 3.11+
# ─────────────────────────────────────────────────────────────

# LangChain ecosystem
langchain>=0.2
langchain-core>=0.2
langchain-groq>=0.1        # Groq LLM integration (Writer, Critic, Ragas judge)
langchain-huggingface>=0.0 # Local BAAI embedding model
langchain-openai>=0.1      # OpenAI-compatible interface (optional fallback)

# LangGraph — multi-agent state machine
langgraph>=0.1

# Ragas — RAG quality evaluation framework (v0.4+)
ragas>=0.1

# HuggingFace datasets (required by Ragas Dataset.from_dict)
datasets>=2.14

# Local sentence embeddings (BAAI/bge-small-en-v1.5, downloaded on first run)
sentence-transformers>=2.2

# ML / numerics
scikit-learn>=1.3
numpy>=1.26
pandas>=2.0

# Environment variable management
python-dotenv>=1.0
